tests/requirements_test.txt
# Test requirements for binary classifier
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
requests-mock>=1.10.0
flask>=2.3.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.3.0
transformers>=4.30.0
pydantic>=2.0.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0